+++ src/renderer/tests/errors.rs
@@ -101,7 +101,7 @@ fn error_out_of_range_index() {
     let mut tera = Tera::default();
     tera.add_raw_templates(vec![("tpl", "{{ arr[10] }}")]).unwrap();
     let mut context = Context::new();
-    context.add("arr", &[1, 2, 3]);
+    context.insert("arr", &[1, 2, 3]);
 
     let result = tera.render("tpl", &Context::new());
 
@@ -116,7 +116,7 @@ fn error_unknown_index_variable() {
     let mut tera = Tera::default();
     tera.add_raw_templates(vec![("tpl", "{{ arr[a] }}")]).unwrap();
     let mut context = Context::new();
-    context.add("arr", &[1, 2, 3]);
+    context.insert("arr", &[1, 2, 3]);
 
     let result = tera.render("tpl", &Context::new());
 
@@ -132,8 +132,8 @@ fn error_invalid_type_index_variable() {
     tera.add_raw_templates(vec![("tpl", "{{ arr[a] }}")]).unwrap();
 
     let mut context = Context::new();
-    context.add("arr", &[1, 2, 3]);
-    context.add("a", &true);
+    context.insert("arr", &[1, 2, 3]);
+    context.insert("a", &true);
 
     let result = tera.render("tpl", &context);