+++ src/tera.rs
@@ -298,7 +298,7 @@ impl Tera {
     /// ```rust,ignore
     /// // Rendering a template with a normal context
     /// let mut context = Context::new();
-    /// context.add("age", 18);
+    /// context.insert("age", 18);
     /// tera.render("hello.html", &context);
     /// // Rendering a template with a struct that impl `Serialize`
     /// tera.render("hello.html", &product);
@@ -330,7 +330,7 @@ impl Tera {
     ///
     /// ```rust,ignore
     /// let mut context = Context::new();
-    /// context.add("greeting", &"hello");
+    /// context.insert("greeting", &"hello");
     /// Tera::one_off("{{ greeting }} world", &context, true);
     /// // Or with a struct that impl Serialize
     /// Tera::one_off("{{ greeting }} world", &user, true);
@@ -588,7 +588,7 @@ impl Tera {
     /// tera.add_raw_template("foo", "\"{{ content }}\"").unwrap();
     /// tera.autoescape_on(vec!["foo"]);
     /// let mut context = Context::new();
-    /// context.add("content", &"Hello\n\'world\"!");
+    /// context.insert("content", &"Hello\n\'world\"!");
     /// let result = tera.render("foo", &context).unwrap();
     /// assert_eq!(result, r#""Hello\n\'world\"!""#);
     ///```
@@ -800,7 +800,7 @@ mod tests {
     #[test]
     fn test_can_autoescape_one_off_template() {
         let mut context = Context::new();
-        context.add("greeting", &"<p>");
+        context.insert("greeting", &"<p>");
         let result = Tera::one_off("{{ greeting }} world", &context, true).unwrap();
 
         assert_eq!(result, "&lt;p&gt; world");
@@ -809,7 +809,7 @@ mod tests {
     #[test]
     fn test_can_disable_autoescape_one_off_template() {
         let mut context = Context::new();
-        context.add("greeting", &"<p>");
+        context.insert("greeting", &"<p>");
         let result = Tera::one_off("{{ greeting }} world", &context, false).unwrap();
 
         assert_eq!(result, "<p> world");
@@ -837,7 +837,7 @@ mod tests {
         tera.autoescape_on(vec!["foo"]);
         tera.set_escape_fn(escape_c_string);
         let mut context = Context::new();
-        context.add("content", &"Hello\n\'world\"!");
+        context.insert("content", &"Hello\n\'world\"!");
         let result = tera.render("foo", &context).unwrap();
         assert_eq!(result, r#""Hello\n\'world\"!""#);
     }
@@ -851,7 +851,7 @@ mod tests {
         tera.set_escape_fn(no_escape);
         tera.reset_escape_fn();
         let mut context = Context::new();
-        context.add("content", &"Hello\n\'world\"!");
+        context.insert("content", &"Hello\n\'world\"!");
         let result = tera.render("foo", &context).unwrap();
         assert_eq!(result, "Hello\n&#x27;world&quot;!");
     }